create internalLinks component
