+++ src/components/InternalLinks/index.ts
@@ -1,0 +1,1 @@
+export * from './internal-links'
