+++ src/components/Navbar/navbar.tsx
@@ -4,22 +4,16 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { ContactButton } from '@/components/ContactButton';
 import lecodeLogo from '../../../public/lecode-logo.svg';
-import { NavbarContainer, NavbarContentContainer, NavLinks, StyledLink } from './styles';
+import { NavbarContainer, NavbarContentContainer} from './styles';
+import { InternalLinks } from '../InternalLinks';
 
 const poppins = Poppins({ weight: ['600'], style: ['normal'], subsets: ['latin'], display: 'swap' });
-const changa = Changa({ weight: ['500'], style: ['normal'], subsets: ['latin'], display: 'swap' });
 
 export const Navbar = () => {
   return (
     <NavbarContainer>
       <NavbarContentContainer>
-        <NavLinks className={changa.className}>
-          <StyledLink href='#about'>Quem Somos</StyledLink>
-          <StyledLink href='#technologies'>Nossas Habilidades</StyledLink>
-          <StyledLink href='#cases'>Cases</StyledLink>
-          <StyledLink href='#team'>Equipe</StyledLink>
-          <StyledLink href='#clients'>Clientes</StyledLink>
-        </NavLinks>
+       <InternalLinks />
         <Image
           alt='LeCode logo: green triangle with white Le and black code'
           src={lecodeLogo}
