+++ src/components/Navbar/styles.tsx
@@ -30,18 +30,4 @@ export const NavbarContentContainer = styled.div`
   width: 90rem;
 `;
 
-export const NavLinks = styled.div`
-  display: flex;
-  padding: 0.625rem;
-  justify-content: center;
-  align-items: center;
-  gap: 1.25rem;
-  color: ${({ theme }) => theme.colors.black};
-  font-size: 1rem;
-`;
 
-export const StyledLink = styled(Link)`
-  &:hover {
-    color: ${({ theme }) => theme.colors.hover.mainGreen};
-  }
-`;
